# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenWrapper Gateway API",
    "version": "0.1.4",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
...
  },
  "servers": [
    {
      "url": "https://gateway.openwrapper.muejam.com",
      "description": "Production OpenWrapper Gateway (TLS - Paymob, Fawry, Stripe)"
    },
    {
      "url": "https://openwrapper.muejam.com/api",
      "description": "Production Web API (Control Plane, Better Auth, & Ledger GraphQL)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Rust gateway (Paymob, Fawry, Stripe, gRPC & GraphQL)"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Web Control Plane & API"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
    "/v1/payments": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create a payment",
        "operationId": "createPayment",
        "description": "Creates one logical payment across the specified provider (Paymob, Fawry, Stripe).\nRequires a caller-supplied `Idempotency-Key`. Retries must reuse both the request\nbody and `Idempotency-Key`. A `200` response represents an idempotent replay or\na recorded payment whose provider outcome is `unknown`. A `201` response indicates\na newly initiated payment intention.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/PaymobSecretKey"
          },
          {
            "$ref": "#/components/parameters/PaymobPublicKey"
          },
          {
            "$ref": "#/components/parameters/PaymobHmacSecret"
          },
          {
            "$ref": "#/components/parameters/PaymobIntegrationId"
          },
          {
            "$ref": "#/components/parameters/PaymobBaseUrl"
          },
          {
            "$ref": "#/components/parameters/FawryMerchantCode"
          },
          {
            "$ref": "#/components/parameters/FawrySecureKey"
          },
          {
            "$ref": "#/components/parameters/FawryBaseUrl"
          },
          {
            "$ref": "#/components/parameters/StripeSecretKey"
          },
          {
            "$ref": "#/components/parameters/StripeWebhookSecret"
          },
          {
            "$ref": "#/components/parameters/StripeBaseUrl"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing payment or recorded ambiguous outcome.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            }
          },
          "201": {
            "description": "Payment initiated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "Idempotency conflict reported by the gateway or web API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request payload or missing provider credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "description": "Required gateway service is not configured or available.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "OpenWrapper API key in `Authorization: Bearer <key>`."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "OpenWrapper API key sent directly in `X-API-Key`."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "1-200 printable ASCII characters, excluding quotes and whitespace.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 200,
          "pattern": "^[!#-~]+$",
          "example": "order-98234-attempt-1"
        }
      },
...
      "PaymobSecretKey": {
        "name": "X-Paymob-Secret-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Per-request Paymob secret API key."
      },
      "PaymobPublicKey": {
        "name": "X-Paymob-Public-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Per-request Paymob public API key."
      },
      "PaymobHmacSecret": {
        "name": "X-Paymob-Hmac-Secret",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Per-request Paymob webhook HMAC secret."
      },
      "PaymobIntegrationId": {
        "name": "X-Paymob-Integration-Id",
        "in": "header",
        "required": false,
        "schema": {
          "oneOf": [
            {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            {
              "type": "integer",
              "format": "int64"
            }
          ]
        },
        "description": "Per-request numeric Paymob integration ID."
      },
      "PaymobBaseUrl": {
        "name": "X-Paymob-Base-Url",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uri"
        },
        "description": "Optional Paymob API origin override for sandbox testing."
      },
      "FawryMerchantCode": {
        "name": "X-Fawry-Merchant-Code",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Per-request Fawry merchant code."
      },
      "FawrySecureKey": {
        "name": "X-Fawry-Secure-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Per-request Fawry shared secret."
      },
      "FawryBaseUrl": {
        "name": "X-Fawry-Base-Url",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uri"
        },
        "description": "Optional Fawry API origin override for sandbox or production."
      },
      "StripeSecretKey": {
        "name": "X-Stripe-Secret-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Per-request Stripe secret API key (e.g. sk_live_... or sk_test_...)."
      },
      "StripeWebhookSecret": {
        "name": "X-Stripe-Webhook-Secret",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Per-request Stripe webhook signing secret (whsec_...)."
      },
      "StripeBaseUrl": {
        "name": "X-Stripe-Base-Url",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uri"
        },
        "description": "Optional Stripe API origin override for sandbox or local mock testing."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request or idempotency payload conflict.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing/invalid API key or rejected provider credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Caller is not authorized for this operation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Gateway or provider rate limit exceeded.",
        "headers": {
          "Retry-After": {
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal gateway failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadGateway": {
        "description": "Provider or downstream gateway failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "GatewayTimeout": {
        "description": "Provider operation exceeded the gateway deadline.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "CreatePaymentRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "provider",
          "amount_minor_units",
          "currency",
          "customer"
        ],
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/PaymentProvider"
          },
          "amount_minor_units": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 1000000000,
            "description": "Integer amount in the currency's minor unit; never a floating-point major-unit value.",
            "example": 15000
          },
          "currency": {
            "$ref": "#/components/schemas/PaymentCurrency"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDetails"
          },
          "merchant_reference": {
            "type": [
              "string",
              "null"
            ],
            "description": "Merchant-owned order or transaction reference.",
            "example": "order-2026-08-31-001"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "Annual Pro Plan Subscription"
          },
          "return_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri",
            "description": "Customer return URL for redirect-based providers."
          },
          "metadata": {
            "type": "object",
            "maxProperties": 20,
            "additionalProperties": {
              "type": "string",
              "maxLength": 500
            },
            "default": {}
          }
        }
      },
...
      "PaymentView": {
        "type": "object",
        "required": [
          "payment_id",
          "provider",
          "provider_reference",
          "status",
          "amount_minor_units",
          "currency",
          "merchant_reference"
        ],
        "properties": {
          "payment_id": {
            "type": "string",
            "minLength": 1
          },
          "provider": {
            "$ref": "#/components/schemas/PaymentProvider"
          },
          "provider_reference": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "amount_minor_units": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 1000000000
          },
          "currency": {
            "$ref": "#/components/schemas/PaymentCurrency"
          },
          "merchant_reference": {
            "type": [
              "string",
              "null"
            ]
          },
          "next_action": {
            "$ref": "#/components/schemas/PaymentNextAction"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
    }
  }
}